Create Dice Class
Dice is the basic randomizer for the RPG set.  A Dice class object is created by assigning Dice(*args) to a variable. The *args are integers and will be split into two groups: d group is a list of all the dice (by number of faces), and n group is the number of the die type just identified. Arguments assigned to Dice() must always be in the form xdy, where x is the number of dice and y is the number of faces on each die of that group. Dice will take any number of dice in a list of strings.